Re-work site and content to use Hugo rendering tmpls
Rework the site to use Hugo as a templating and rendering engine. Update
README.md to reflect this change and document new install instructions.

Documenting releases, adding maintainers, and adjusting community links
/ resources should be simpler now that Hugo templates are handling much
of the rendering logic.

I've kept much of the branding the same but the underlying HTML and
styling is far different than before. Still mobile-friendly!

Signed-off-by: Wesley Coakley <wcoakley@nvidia.com>